requirements.txt
Flask
requests
python-dotenv
gspread
oauth2client
sendgrid
firebase-admin
google-cloud-bigquery
